refactor: Update middleware, program configuration, and repository interface

+++ server.NET/OracleMES.Core/Interfaces/Repositories/IRepository.cs
@@ -5,7 +5,7 @@ namespace OracleMES.Core.Interfaces.Repositories;
 public interface IRepository<T> where T : class
 {
     Task<IEnumerable<T>> GetAllAsync();
-    Task<T?> GetByIdAsync(int id);
+    Task<T?> GetByIdAsync(decimal id);
     Task<T> AddAsync(T entity);
     Task<IEnumerable<T>> AddRangeAsync(IEnumerable<T> entities);
     Task<T> UpdateAsync(T entity);
